Avance 3

+++ app/Pedido.php
@@ -38,6 +38,15 @@ class Pedido extends Model
         return collect( Pedido::where('ciclo',$id_ciclo)->where('tipo',1)->get() );
     }
 
+    public static function limpiaErroresBBDD($iteraciones = 6){
+        for ($x=0;$x<$iteraciones;$x++){
+            self::eliminaDuplicados();
+            self::diasAbsorbenJornadasMismoTipoMismoDia();
+            self::mañanaTardesMismoTipoMismoDia();
+            self::dosDiasYNocheIgualCiclo();
+        }
+    }
+
     public static function mañanaTardesMismoTipoMismoDia(){
         $sql = "SELECT p1.id,p2.id as id2 FROM pedidos as p1,pedidos as p2 WHERE ";
         $sql .= "p1.ciclo=p2.ciclo AND p1.tipo=p2.tipo AND ";
@@ -47,11 +56,14 @@ class Pedido extends Model
             if (isset($par_de_ids->id) and isset($par_de_ids->id2)) {
                 $pedido_temp1 = Pedido::find($par_de_ids->id);
                 $pedido_temp2 = Pedido::find($par_de_ids->id2);
-                $pedidoAMeter = $pedido_temp1->replicate();
-                $pedidoAMeter->que_pide = 1;
-                $pedidoAMeter->save();
-                $pedido_temp1->delete();
-                $pedido_temp2->delete();
+                if ($pedido_temp1!=null and $pedido_temp2!=null){
+                    $pedidoAMeter = $pedido_temp1->replicate();
+                    $pedidoAMeter->push();
+                    $pedidoAMeter->que_pide = 1;
+                    $pedidoAMeter->save();
+                    $pedido_temp1->delete();
+                    $pedido_temp2->delete();
+                }
             }
         }
 
@@ -63,12 +75,112 @@ class Pedido extends Model
             if (isset($par_de_ids->id) and isset($par_de_ids->id2)) {
                 $pedido_temp1 = Pedido::find($par_de_ids->id);
                 $pedido_temp2 = Pedido::find($par_de_ids->id2);
-                $pedidoAMeter = $pedido_temp1->replicate();
-                $pedidoAMeter->que_pide = 2;
-                $pedidoAMeter->save();
-                $pedido_temp1->delete();
-                $pedido_temp2->delete();
+                if ($pedido_temp1!=null and $pedido_temp2!=null) {
+                    $pedidoAMeter = $pedido_temp1->replicate();
+                    $pedidoAMeter->push();
+                    $pedidoAMeter->que_pide = 2;
+                    $pedidoAMeter->save();
+                    $pedido_temp1->delete();
+                    $pedido_temp2->delete();
+                }
             }
         }
+
+        return true;
+    }
+
+    public static function diasAbsorbenJornadasMismoTipoMismoDia(){
+        $sql = "SELECT p1.id,p2.id as id2 FROM pedidos as p1,pedidos as p2 WHERE ";
+        $sql .= "p1.ciclo=p2.ciclo AND p1.tipo=p2.tipo AND ";
+        $sql .= "(p1.que_pide=1 AND (p2.que_pide=4 OR p2.que_pide=3))";
+        $temp = DB::select($sql);
+        foreach($temp as $par_de_ids){
+            if (isset($par_de_ids->id) and isset($par_de_ids->id2)) {
+                $pedido_temp1 = Pedido::find($par_de_ids->id);
+                $pedido_temp2 = Pedido::find($par_de_ids->id2);
+                if ($pedido_temp1!=null and $pedido_temp2!=null){
+                    $pedido_temp2->delete();
+                }
+            }
+        }
+
+        $sql = "SELECT p1.id,p2.id as id2 FROM pedidos as p1,pedidos as p2 WHERE ";
+        $sql .= "p1.ciclo=p2.ciclo AND p1.tipo=p2.tipo AND ";
+        $sql .= "(p1.que_pide=2 AND (p2.que_pide=5 OR p2.que_pide=6))";
+        $temp = DB::select($sql);
+        foreach($temp as $par_de_ids){
+            if (isset($par_de_ids->id) and isset($par_de_ids->id2)) {
+                $pedido_temp1 = Pedido::find($par_de_ids->id);
+                $pedido_temp2 = Pedido::find($par_de_ids->id2);
+                if ($pedido_temp1!=null and $pedido_temp2!=null){
+                    $pedido_temp2->delete();
+                }
+            }
+        }
+
+        return true;
+    }
+
+    public static function eliminaDuplicados(){
+        $sql = "SELECT p1.id,p2.id as id2 FROM pedidos as p1,pedidos as p2 WHERE ";
+        $sql .= "p1.ciclo=p2.ciclo AND p1.id!=p2.id AND ";
+        $sql .= "p1.que_pide=p2.que_pide AND p1.user_id=p2.user_id";
+    //    dd($sql);
+        $temp = DB::select($sql);
+        $x = 0;
+        $y = 0;
+        foreach($temp as $par_de_ids){
+            if (isset($par_de_ids->id) and isset($par_de_ids->id2)) {
+                $pedido_temp1 = Pedido::find($par_de_ids->id);
+                $pedido_temp2 = Pedido::find($par_de_ids->id2);
+                if ($pedido_temp1!=null and $pedido_temp2!=null){
+                    $pedidoAMeter = $pedido_temp1->replicate();
+                    $pedidoAMeter->push();
+                    $pedidoAMeter->que_pide = 1;
+                    if ($pedidoAMeter->save())
+                        $y++;
+                    if ($pedido_temp1->delete())
+                        $x++;
+                    if ($pedido_temp2->delete())
+                        $x++;
+
+                }
+            }
+        }
+        echo "$x BORRADOS, $y GUARDADOS<br>";
+        return true;
+    }
+
+    public static function dosDiasYNocheIgualCiclo(){
+        $sql = "SELECT p1.id,p1.que_pide,p2.id as id2,p2.que_pide,p3.id as id3,p3.que_pide,p1.ciclo,p2.ciclo,p3.ciclo ";
+        $sql .= "FROM pedidos as p1,pedidos as p2, pedidos as p3 WHERE ";
+        $sql .= "p1.ciclo=p2.ciclo AND p2.ciclo=p3.ciclo AND p1.id!=p2.id AND p2.id!=p3.id AND ";
+        $sql .= "p1.que_pide=1 AND p2.que_pide=2 AND p3.que_pide=7";
+
+        $temp = DB::select($sql);
+        $x = 0;
+        $y = 0;
+        foreach($temp as $par_de_ids){
+            if (isset($par_de_ids->id) and isset($par_de_ids->id2) and isset($par_de_ids->id3)) {
+                $pedido_temp1 = Pedido::find($par_de_ids->id);
+                $pedido_temp2 = Pedido::find($par_de_ids->id2);
+                $pedido_temp3 = Pedido::find($par_de_ids->id3);
+                if ($pedido_temp1!=null and $pedido_temp2!=null and $pedido_temp3!=null){
+                    $pedidoAMeter = $pedido_temp1->replicate();
+                    $pedidoAMeter->push();
+                    $pedidoAMeter->que_pide = 0;
+                    if ($pedidoAMeter->save())
+                        $y++;
+                    if ($pedido_temp1->delete())
+                        $x++;
+                    if ($pedido_temp2->delete())
+                        $x++;
+                    if ($pedido_temp3->delete())
+                        $x++;
+                }
+            }
+        }
+        echo "$x BORRADOS, $y GUARDADOS<br>";
+        return true;
     }
 }
